Revert "Declare previous version when using frozen HALs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3df9b93e-5a29-49c8-970f-7c44c604fc21). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I602a1a4440227094b6fa0a2f284cee231f9f42d8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com