Revert "Set input thread priority to RT"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bacd64c4-6d38-4789-bbf6-cc5cc636fea8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I88c9d85f8e66dc117d3491221b16664f89662456
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com