Revert "Add source of unwind when fatal error."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e5356f01-091e-4434-9f83-4a484ba979df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1b9088f525246230b47e47a3416a6c26cb6dde97
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com