Revert "Fix failure in CowTest#InvalidMergeOrderTest."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e5356f01-091e-4434-9f83-4a484ba979df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If08bf2835779717db467e1eaf2b16a05f5a01a2d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com