Revert "Fix failure in CowTest#InvalidMergeOrderTest."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/194605cb-57ad-4595-85a5-4a4332cced8c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia4a7aa74dc855885476cd2d6126632111dfb399d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com