Revert "Add source of unwind when fatal error."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/194605cb-57ad-4595-85a5-4a4332cced8c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I97b5419844eeaf67268bc80e0de26e868c1168ec
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com