Revert "trusty: utils: rpmb_dev: add wv secure storage init.rc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd3e9109-598b-4529-983f-df3d89ad68d2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic05d459d3571ff34ba67c1b5a5ba45fc193cdf01
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com