Revert "trusty: utils: rpmb_dev: add wv secure storage init.rc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d87ad3cb-2e93-4e38-b27c-b20fc82d96d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I10009b00e33b9f6227a63b05da41e0461acf3ec3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com