Revert "Add source of unwind when fatal error."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0efe22f5-3369-4473-a524-ec667e8aa5d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic7930c8ad862dc6da1f5a725f71380f3c9d7621d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com