Revert "Fix failure in CowTest#InvalidMergeOrderTest."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/389dacc6-b91f-4c88-b3e8-ac967229a399). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieec83bd10b646c428c206c3ed6835dcb816ed89b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com