Revert "trusty: utils: rpmb_dev: add wv secure storage init.rc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cc301e77-cb60-4a28-8255-02449a33b755). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7df57802409900e521a17efe7b045ec3da33189d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com