Revert "trusty: utils: rpmb_dev: add wv secure storage init.rc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/00767f3a-e6ff-42d9-9c3c-6ffa5437496e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7ca8b833e865a97c639cf20b85541c068694dda4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com