Revert "Increase the test timeout"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/78a596f9-4215-4609-8c5f-cf4088f8320a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I75ff3b6c95cfcc1bc0c7b9e28bbaec1653cd1960
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com