watchdogd: Switch to high-pri SCHED_FIFO
Summary: We're seeing cases where we're getting kernel panics from the softdogbiting, and this seems to be mostly during high loads/heavy lmkdactivity. Schedule the watchdog petting thread SCHED_FIFO and at highpriority to (hopefully) eliminate the possibility that scheduling getsblocked and we miss it. Note that we also need to allow this schedulingchange in sepolicy, which I'll do in a separate diff.

Test: Successful Build on master branch

Change-Id: Idb9cfdaf2ef53804ae8185658811a96ad00a3943
Signed-off-by: Abdelrahman Daim <adaim@meta.com>