Revert "Reapply "libprocessgroup: Remove __BEGIN_DECLS and __END..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0e7b25b1-52c9-4964-87fa-6107c03b3e9f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id3438dfa5ea575feb7ca091b8f76f2f160c76fe8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com