Revert "libprefetch: Start prefetch service based on build"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7fcec49f-c5ea-4f1a-b187-0675ec12b89c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id561184a99307b70cb4f8acdae12228f1c639d52
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com