Revert "move aconfigd platform init service from init.rc into ac..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c75e3df-428e-40a9-92e8-da04e691804d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I438184088447a11489e58187ba4cd2c6471b084b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com