Revert "move aconfigd platform init service from init.rc into ac..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1c21271-8c9f-434f-b1fb-98b5376e797a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If97412009fc46a4383a6bc0521350c8b1f3296f3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com