Revert "trusty: utils: rpmb_dev: add wv secure storage init.rc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9dea3dd-ed12-4fb0-8683-4ac884e86c90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I17636fb75cd08778ed241a16646148c61c165d82
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com