Revert "move aconfigd platform init service from init.rc into ac..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9ccf9e4a-7915-4830-87fb-19f7dbd8574a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I35268b2c847c9a642bb5f2c91d734f305abd36a8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com