Revert "trusty: utils: rpmb_dev: add wv secure storage init.rc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe43564e-9fee-4b93-9c1c-185bacda4ce0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iebf3d3b0d836e4e7e9ba39a2ffbb05e335b8f6e2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com