Revert "libprefetch: Start prefetch service based on build"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fdedaab2-7342-4340-9566-ddcf1d4096a3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4f6a3d24170bcdda6b3fdacc40d9511ce5b7527f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com