Revert "move aconfigd platform init service from init.rc into ac..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e31a1a4-9859-484b-a1de-076faad90425). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ief9937f5df18941b439841f0608625f43abbe3de
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com