Revert "move aconfigd platform init service from init.rc into ac..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ce87a0ec-22b4-48ed-9f90-07f78616629c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I29c0fbffe114bf93e04494b858ab7eb60c90e438
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com