Revert "move aconfigd platform init service from init.rc into ac..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a9e3b163-09cb-46e8-ac8a-1fc450a23de3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I094b5f71d8923c7ea63376f8e66920245b9ea69e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com