Revert "move aconfigd platform init service from init.rc into ac..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dae0a3ec-6e0e-4591-b3aa-19afc192a2b3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I76b8c8ce6415a6b68712d7810b981f6d86cd3bce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com