Revert "Update trusty to use secretkeeper hal V1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8c18286-cfdf-4234-98b5-5b6d2795c1f2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 382376014
Change-Id: If3dd27354a59a0d98ae3102bc7891e2eea7f685d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com