Revert "ashmem: Ensure all memfds have non-executable permission..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/15a42795-7e7e-46fd-8a1a-65157c93f628). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d07ef75ee45d13f7d9ab5e6e3a045b6d35b85ac
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com