Revert "Define ueventd.rc.recovery"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb7e03c4-d153-4b95-8ef2-d7fb6c001029). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 382795480
Change-Id: I4c8867c4bf7b45483369587b91360e5c8aff3e44
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com