Revert "Define ueventd.rc.recovery"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b0b82c97-9e02-42b5-9f0f-c3d23ff02916). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383033001
Change-Id: Ic428a5209e8e6fdfbd506705a589d58447f82cc5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com