Revert "Add support for getting TLS on all arches."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e95bd1a9-fc4f-4f76-a8a5-de45ad3a61d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I85f5d90686a61fd8db7256e26595d7264daf560f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com