Revert "Add support for getting TLS on all arches."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1148344-7c73-44e3-af70-54bb32cd768f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6106c45d9ddbe37b45d1d2e8b3b7d955e6236cdd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com