Revert "Add support for getting TLS on all arches."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/60c7fdb4-cae6-45d1-82ca-a605d079bfe3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1cd42dc9b78de30c586c474e456a2b085a5ea490
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com