Revert "Add support for getting TLS on all arches."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8881a26e-6ba8-4681-b2eb-b2dfeafa238f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I429351721ea0b344216a469e5fbc69e878fd8110
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com