Revert "Update trusty to use secretkeeper hal V1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7fc9b1de-6773-4f88-98c3-ae90b94e5598). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383543807
Change-Id: I452245b10e4517224edc1a5af13a263bb8349940
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com