Revert "Add support for getting TLS on all arches."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6f9f4f57-05dd-43a5-86a7-54dd0a2ec37f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0baf9f1b50418d91c2344d1a00564bed1f0d745b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com