Revert "Define reboot.recovery and watchdogd.recovery"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ca868fb-e51c-470f-b6ac-3860b904449c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d0252ef7e14d006dde2dd7ac2e4a10adf916e4a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com