Revert "Add support for getting TLS on all arches."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/001703c5-ab7b-4c5f-b8be-d88b7dec8ef1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibcee61c2ac63b6a60e08d22ac7615a970f62b661
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com