Revert "Add support for getting TLS on all arches."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21330848-afe2-4faf-9ff7-1e5ee61be818). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0e5af779f9fbfa1e8e57f7fc438c98b40196a8ed
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com