Revert "libprocessgroup: memcg default version 1 -> 2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/84569e33-a91e-45ed-bc0f-c0929649ced1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 384285412
Bug: 384285412
Change-Id: I4a0f42b80f2cd51c32acd6b4bb86cd545b856a66
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com