Revert "libprocessgroup: Explicit field presence for MaxActivati..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/84569e33-a91e-45ed-bc0f-c0929649ced1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 384285412
Bug: 384285412
Change-Id: If504d32b410e027198e99a2b41f4feb9e7022d9a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com