libprefetch: fix property names in the .rc script
Current property names in .rc do not match what libprefetch actually
expects. In addition, the name of the property to stop the record starts
with `ro.` while it is expected to be modified. This commit fixes these
problems. Since prefetch was not working due to mismatching names, this
fix should not introduce a breaking change.

Bug: 362507272
Bug: 380766679
Test: Test prefetch works on a Cuttlefish
Change-Id: I588b6f555fa9797a1a089aa983bcdc47bbb5e17e
Signed-off-by: Takaya Saeki <takayas@google.com>