Revert "Define reboot.recovery and watchdogd.recovery"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/67f1c5cf-fdd8-444b-8bc6-6d1978b48014). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id827c8ec83957b6eb270b8daf49d40867974f378
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com